Title: 10개 입력받기, Time: 117ms, Memory: 10MB, Status: Passed - Codetree

+++ 251222/10개 입력받기/receive-10-inputs.java
@@ -14,7 +14,7 @@ public class Main {
             cnt++;
         }
 
-        double avg = Math.round(sum / cnt * 10.0) / 10.0;
+        double avg = Math.round((double) sum / cnt * 10.0) / 10.0;
 
         System.out.println(sum + " " + avg);
     }
